Initial image support
This is first version with image fetching support. Currently, it lazy
loads the images and renders them as ansi. A configuration window is
added where you can select if images should be rendered (default off)
and what render mechanism it should use. Sixel is in the list, but
currently it just inserts a placeholder if that is selected.