user selectable dark and light theme
Both the light and the dark theme selection is nwo stored in the config
file. If the user selects a theme with "light" in the name, it will get
stored as preferred light theme. If the user selects another theme, then
it will be stored as preferred dark theme. The user can toggle between
them by pressing d on the main screen.