feat(Caching): Added caching db for backend
The system now uses an sqlite db as backend and uses it for caching of
messages. Images will be cached in a folder. The app will load messages
and images from the cache if available and only fetch new, uncached messages.
If the messages in the cache are more than 2 hours old, it will start to
fetch from the server and only when the gap is 30mins, will it try to
fetch the remaining messages to fill the gap.
Automatic pruning of the chache happens on startup and will remove all
items older than 30 days.