feat(UI): custom theme support, clickable links/hashtags in timelines
This was a bit of a messy change but it was preperations for the bit
1.0.0 release. I have added custom theme support so the user can create
themes as they like them and select them via the ctrl-p menu.
I also made it possible to click on hashtags and links in the timelines.
If a hashtag is clicked, then a hashtag timeline is shown.
A lot of UI rendering has been simplified, cleaned up and optimized.
Support for "update" notifications have also been added.
All in all a rather big commit this time as I forgot to split them in
time.