feat(direct messages): post, reply direct messages, new message notif
Made it possible to to reply to direct messages where the visibility of
the reply is set correctly (same for other replies that now follow the
original message visibility).
Added a "new direct message" icon to the upper right corner if a new
direct message arrives while the direct messages timeline is hidden
(easily toggeled with m key).
Code cleanup and some refactoring due to a bit too aggressive cleanup
earlier.