fix(issue#11): possible fix for noneType error
The report in issue#11 did not give much info about where in the mastui
code this happened, but from the info I had, I think I have found the
problem. I believe that this is cause by a malformed poll (maybe one
that was deleted and then ended). This fix should handle that
situation.