docs(ROADMAP): Added a roadmap file
I have many ideas and requests on my todo-list. Wanted to get them all
into a file for others to read as well. The file has some initial
architecture info and then the roadmap to make it easier for others to
contribute.
README was also updated with a few missing features.