feat(autocomplete): initial auto complete support
My first stab at adding user and hashtag completion. start with @ or #
and minimum 2 characters, then press tab. A list will show (it can be
slow right now). Use tab or ctrl-n to move down the list or ctrl-p to go
up the list. Enter to insert the selected item or escape to do nothing.